meesho-guide-ai/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
langchain==0.3.27
sentence-transformers==5.1.0
pinecone-client==6.0.0
python-dotenv==1.1.1
websockets==15.0.1
pydantic==2.11.9
pydantic-settings==2.10.1
